Perform matching between all TF Algorithms
This code now does not only estimate the BMTF efficiency but also the
OMTF and EMTF and the combined uGMT(i.e. all Tf Algorithms over the
whole eta range).
The Tag and Probe method was now also modfiied.
Now multiple tags can be identified.
For each tag the alorithm tries to find another reco muon for which
the dimuon mass is close to the zBoson or the jPsi.
The reco muon that is closest to either mass will be the probe muon.